Prototype of the core tiling algorithm.
Note that in this commit the tiling algorithm is committed, and the glue to use it in Motile is there, but it is not yet integrated into the application because the colors are not working yet. Once they are working it will be part of the core UI, etc etc.